% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotter.R
\name{color_agents}
\alias{color_agents}
\title{Color agents based on values}
\usage{
color_agents(values, col_palette = NULL)
}
\arguments{
\item{values}{Numeric vector of values to map to colors.}

\item{col_palette}{Optional character vector of colors. If NULL, a blue-white-red palette of 100 colors is used.}
}
\value{
Character vector of colors corresponding to the input values.
}
\description{
Maps numeric values to colors using a color palette.
}
