% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neighbors.R
\name{create_neighborhood}
\alias{create_neighborhood}
\title{Create a neighborhood based on influence patterns}
\usage{
create_neighborhood(agents, neighbors, params = list())
}
\arguments{
\item{agents}{A data frame of agents, typically created by \code{create_agents}.}

\item{neighbors}{A function that returns the neighbors of a given agent.}

\item{params}{A list of additional parameters passed to the neighbors function.}
}
\value{
A list where each element contains the neighbors of the corresponding agent.
}
\description{
Generates a list of neighbors for each agent according to a specified neighbor function.
}
