% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsitR-package.R
\docType{package}
\name{dsitR}
\alias{dsitR}
\alias{dsitR-package}
\title{dsitR: Dynamic Social Impact Simulator}
\description{
\code{dsitR} is an R package for simulating Dynamic Social Impact Theory (DSIT).
It models agents with multi-dimensional opinions who influence each other through customizable neighborhoods.
The package includes modular tools for simulation, metrics, and visualization of cultural dynamics over time.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cwendorf/dsitR}
  \item Report bugs at \url{https://github.com/cwendorf/dsitR/issues}
}

}
\author{
\strong{Maintainer}: Craig Wendorf
\email{cwendorf@uwsp.edu}
}
