% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neighbors.R
\name{neighbors_cluster}
\alias{neighbors_cluster}
\title{Cluster neighbors (same row)}
\usage{
neighbors_cluster(id, agents, params = list())
}
\arguments{
\item{id}{Integer ID of the agent.}

\item{agents}{A data frame of agents with \code{x} and \code{y} positions.}

\item{params}{A list of additional parameters. \code{params$cluster_size} specifies the number of neighbors to select (default 3).}
}
\value{
Integer vector of neighbor agent IDs from the same row.
}
\description{
Returns the indices of neighbors in the same row (cluster) as the agent.
}
