% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neighbors.R
\name{neighbors_random_k}
\alias{neighbors_random_k}
\title{Random k neighbors}
\usage{
neighbors_random_k(id, agents, params = list())
}
\arguments{
\item{id}{Integer ID of the agent.}

\item{agents}{A data frame of agents.}

\item{params}{A list of additional parameters. \code{params$k} specifies the number of neighbors (default 3).}
}
\value{
Integer vector of randomly selected neighbor agent IDs.
}
\description{
Returns the indices of k randomly selected neighbors for an agent,
excluding the agent itself.
}
